+++ docs/src/examples/templates/template-event-handlers/component.js
@@ -1,0 +1,23 @@
+import { defineComponent, getText } from '@semantic-ui/component';
+
+const css = await getText('./component.css');
+const template = await getText('./component.html');
+
+const defaultState = {
+  checked: false
+};
+
+const createComponent = ({ state }) => ({
+  toggleCheckbox() {
+    state.checked.toggle();
+  }
+});
+
+
+defineComponent({
+  tagName: 'template-events',
+  template,
+  css,
+  defaultState,
+  createComponent,
+});