+++ RELEASE-NOTES.md
@@ -11,6 +11,8 @@ Please note after `1.0` Semver will be followed using normal protocols.
 ## Templates
 * **Feature** - Added template helpers: `default`, `truncate`, `first`, `last`, `roundNumber`, `roundDecimal` with `round` alias, `lowercase`, `uppercase`
 * **Feature** - Added `registerHelper()` and `registerHelpers()` functions for registering custom template helpers
+* **Features* - Added support for binding events from inside templates using `@` handlers like `<div @click={doSomething}></div>`
+* **Feature** - Added support for binding el properties from inside templates like `<input type="checkbox" .checked={checked}>`
 
 ## Utils
 * **Feature** - Added `truncate` utility function to utils package for word-boundary aware text truncation